Added tower selecting and test scene artwork
Can now select towers and their radius will be displayed. Once more functionality is added it will also show, such as current upgrades or even tower stats (range, attack damage, health, etc).

Test scene artwork. This is mainly for testing. We need to add a collider layer to the tile map to now allow the tower to be placed on the track. Will work on that next.